pyproject.toml
[tool.poetry]
name = "classcom"
version = "0.1.0"
description = ""
authors = ["A'zamov Samandar <azamov.samandar.programmer@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 79

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = "tests.py test_*.py *_tests.py"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::ResourceWarning",
    "ignore::Warning" # This line will ignore all warnings
]
addopts = "--cov=config --cov-report=term"

[tool.poetry.dependencies]
python = "^3.10"
cookiecutter = "2.5.0"
crispy-tailwind = "1.0.1"
django-ckeditor = "6.7.1"
django-ckeditor-5 = "0.2.11"
django-cors-headers = "4.3.1"
django-cron = "0.6.0"
django-debug-toolbar = "4.3.0"
django-environ = "0.11.2"
django-extensions = "3.2.3"
django-filter = "23.5"
django-import-export = "3.3.7"
django-jazzmin = "3.0.0"
django-modeltranslation = "0.18.11"
django-polymorphic = "3.1.0"
django-redis = "5.4.0"
django-rosetta = "0.10.0"
django-select2 = "8.1.2"
django-unicorn = "0.58.1"
djangorestframework-simplejwt = "5.3.1"
faker = "23.1.0"
flower = "2.0.1"
ipython = "8.21.0"
markuppy = "1.14"
ngrok = "1.2.0"
odfpy = "1.4.1"
openpyxl = "3.1.2"
pip-chill = "1.0.3"
psycopg2-binary = "2.9.9"
pytest = "8.0.0"
shellingham = "1.5.4"
tqdm = "4.66.2"
twine = "4.0.2"
typer = "0.12.0"
xlrd = "2.0.1"
xlwt = "1.3.0"
black = "^24.4.2"
drf-spectacular = "^0.27.2"
pydantic = "^2.7.1"
gunicorn = "^22.0.0"
pandas = "^2.2.2"
pytest-django = "^4.8.0"
coverage = "^7.5.4"
pytest-cov = "^5.0.0"
uzum-payments = "^0.0.6"
django-unfold = "^0.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
